# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Halal Grocery Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores/{storeSlug}/products:
    get:
      operationId: getStoreProducts
      tags: [stores]
      summary: Get products for a store
      parameters:
        - name: storeSlug
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: productType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 40
      responses:
        "200":
          description: Store products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List/search products across all stores
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: productType
          in: query
          schema:
            type: string
        - name: isFreshMeat
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 40
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /products/{productSlug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail by slug
      parameters:
        - name: productSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ORDERS
...
  /store-portal/products:
    get:
      operationId: getStorePortalProducts
      tags: [products]
      summary: Get store's products for management
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Store products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /store-portal/products/{productId}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Store:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        logo:
          type: string
          nullable: true
        banner:
          type: string
          nullable: true
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        lat:
          type: number
        lng:
          type: number
        isHalalCertified:
          type: boolean
        halalCertNumber:
          type: string
          nullable: true
        rating:
          type: number
        totalRatings:
          type: integer
        avgPrepTimeMinutes:
          type: integer
        deliveryAvailable:
          type: boolean
        pickupAvailable:
          type: boolean
        curbsideAvailable:
          type: boolean
        minOrderAmount:
          type: number
        convenienceFee:
          type: number
        openingHoursJson:
          type: object
        isMasjidAffiliated:
          type: boolean
        masjidName:
          type: string
          nullable: true
        isActive:
          type: boolean
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - phone
        - address
        - city
        - province
        - postalCode
        - lat
        - lng
        - isHalalCertified
        - rating
        - totalRatings
        - avgPrepTimeMinutes
        - deliveryAvailable
        - pickupAvailable
        - curbsideAvailable
        - minOrderAmount
        - convenienceFee
        - openingHoursJson
        - isMasjidAffiliated
        - isActive
        - isApproved
        - createdAt
...
    Product:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        productType:
          type: string
          enum: [PACKAGED, FRESH_MEAT, PRODUCE, FROZEN, SPICES, BAKERY, DAIRY, BEVERAGES]
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        price:
          type: number
        comparePrice:
          type: number
          nullable: true
        unit:
          type: string
        stockQty:
          type: number
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        isHalalCertified:
          type: boolean
        isFreshMeat:
          type: boolean
        meatAnimalType:
          type: string
          nullable: true
        estimatedWeightKg:
          type: number
          nullable: true
        pricePerKg:
          type: number
          nullable: true
        tags:
          type: array
          items:
            type: string
        freshnessLabel:
          type: string
          nullable: true
        store:
          $ref: "#/components/schemas/Store"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - slug
        - name
        - productType
        - category
        - images
        - price
        - unit
        - stockQty
        - isActive
        - isFeatured
        - isHalalCertified
        - isFreshMeat
        - tags
        - createdAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
...
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        productType:
          type: string
          enum: [PACKAGED, FRESH_MEAT, PRODUCE, FROZEN, SPICES, BAKERY, DAIRY, BEVERAGES]
        category:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        price:
          type: number
        comparePrice:
          type: number
        unit:
          type: string
        stockQty:
          type: number
        isFreshMeat:
          type: boolean
        meatAnimalType:
          type: string
        pricePerKg:
          type: number
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
      required:
        - name
        - productType
        - category
        - price
        - unit
        - stockQty
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        stockQty:
          type: number
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string

    # ORDER SCHEMAS
...